requirements.txt
usearch>=2.8.12
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
numpy>=1.24.0
langchain>=0.0.335
python-docx>=1.0.0
docx2txt>=0.8
pypdf>=3.17.1
sentence-transformers>=2.2.2
anthropic>=0.8.1
